move signature stuff to its own file
